@@ -77,26 +82,13 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<encoding>UTF-8</encoding>
-		<maven.build.timestamp.format>dd/MM/yyyy HH:mm</maven.build.timestamp.format>
-		<java.version>1.8</java.version>
-				
-		<hibernate.version>5.4.24.Final</hibernate.version>
-		<slf4j.version>1.7.30</slf4j.version>
-		<logback.version>1.2.11</logback.version>
+		<java.version>17</java.version>
+		<maven.compiler.release.version>17</maven.compiler.release.version>
 		
-		<spring.version>5.3.20</spring.version>
-		<spring.security.version>5.5.7</spring.security.version>
-		<aspectj.version>1.9.5</aspectj.version>
 		<javax.inject.version>1</javax.inject.version>
-		<jackson.version>2.12.7</jackson.version>
-		<jms.version>2.0.1</jms.version>
 		<commons.codec.version>1.9</commons.codec.version>
 		<java-jwt.version>3.4.1</java-jwt.version>
 		<bouncycastle.version>1.61</bouncycastle.version>
-		<servlet.version>3.1.0</servlet.version>
-		<jsf.version>2.2.13</jsf.version>
-		<primefaces.version>8.0</primefaces.version>
-		<junit.version>4.13.1</junit.version>
 	</properties>
 
 	<dependencyManagement>
