@@ -1,8 +1,7 @@
 @XmlSchema(namespace = "http://commons.sklsft.org/schema/model/dummies",
 xmlns = {@XmlNs( prefix = "metadata", namespaceURI = "http://commons.sklsft.org/schema/model/dummies") },
-elementFormDefault = javax.xml.bind.annotation.XmlNsForm.UNSET)
+elementFormDefault = jakarta.xml.bind.annotation.XmlNsForm.UNSET)
 package org.sklsft.commons.text.miscellanous;
 
-import javax.xml.bind.annotation.XmlNs;
-import javax.xml.bind.annotation.XmlSchema;
-
+import jakarta.xml.bind.annotation.XmlNs;
+import jakarta.xml.bind.annotation.XmlSchema;
